add: share button on modal
