+++ components/gallery/GalleryItemButton/GalleryItemShareBtn.vue
@@ -1,66 +1,9 @@
 <template>
   <div>
-    <NeoDropdown position="bottom-left" :mobile-modal="false">
-      <template #trigger="{ active }">
-        <NeoButton
-          :class="isMobileDevice ? 'icon-action' : ''"
-          :label="label"
-          :icon="icon"
-          :active="active" />
-      </template>
-
-      <NeoDropdownItem
-        v-clipboard:copy="realworldFullPathShare"
-        @click="toast(String($t('toast.urlCopy')))">
-        Copy Link
-      </NeoDropdownItem>
-      <NeoDropdownItem @click="isModalActive = true">QR Code</NeoDropdownItem>
-      <NeoDropdownItem
-        data-testid="gallery-item-share-dropdown-twitter"
-        @click="actionTwitterShare">
-        Share On Twitter
-      </NeoDropdownItem>
-    </NeoDropdown>
-
-    <NeoModal :value="isModalActive" @close="isModalActive = false">
-      <div class="card">
-        <header class="card-header">
-          <p class="card-header-title">QR Code</p>
-        </header>
-        <div class="card-content">
-          <QRCode
-            :text="realworldFullPathShare"
-            data-testid="gallery-item-share-dropdown-qrcode" />
-        </div>
-      </div>
-    </NeoModal>
+    <ShareButton />
   </div>
 </template>
 
 <script setup lang="ts">
-import {
-  NeoButton,
-  NeoDropdown,
-  NeoDropdownItem,
-  NeoModal,
-} from '@kodadot1/brick'
-import { isMobileDevice } from '@/utils/extension'
-import QRCode from '@/components/shared/QRCode.vue'
-
-const route = useRoute()
-const { $i18n } = useNuxtApp()
-const { toast } = useToast()
-
-const { shareOnX } = useSocialShare()
-
-const isModalActive = ref(false)
-const sharingTxt = $i18n.t('sharing.nft')
-const realworldFullPathShare = ref(`${window.location.origin}${route.fullPath}`)
-
-const label = computed(() => (isMobileDevice ? '' : 'Share'))
-const icon = computed(() => (isMobileDevice ? 'share' : ''))
-
-const actionTwitterShare = (): void => {
-  shareOnX(sharingTxt, realworldFullPathShare.value)
-}
+import ShareButton from '@/components/common/ShareButton.vue'
 </script>
